+++ app/builders/dnd2024_character/backgrounds/scribe_builder.rb
@@ -5,6 +5,8 @@ module Dnd2024Character
     class ScribeBuilder
       def call(result:)
         result[:selected_skills] = { investigation: 1, perception: 1 }
+        result[:ability_boosts] = %w[dex int wis]
+        result[:any_skill_boosts] += 3
 
         result
       end