+++ app/builders/dnd2024_character/backgrounds/soldier_builder.rb
@@ -6,6 +6,7 @@ module Dnd2024Character
       def call(result:)
         result[:selected_feats] = ['savage_attacker']
         result[:selected_skills] = { athletics: 1, intimidation: 1 }
+        result[:ability_boosts] = %w[str dex con]
 
         result
       end