+++ app/builders/dnd2024_character/classes/barbarian_builder.rb
@@ -9,8 +9,10 @@ module Dnd2024Character
       def call(result:)
         result[:weapon_core_skills] = result[:weapon_core_skills].concat(WEAPON_CORE).uniq
         result[:armor_proficiency] = result[:armor_proficiency].concat(ARMOR).uniq
-        result[:abilities] = { str: 15, dex: 13, con: 14, int: 10, wis: 11, cha: 12 }
+        result[:abilities] = { str: 15, dex: 13, con: 14, int: 8, wis: 12, cha: 10 }
         result[:health] = { current: 14, max: 14, temp: 0 }
+        result[:skill_boosts] += 2
+        result[:skill_boosts_list] = %w[animal athletics nature intimidation perception survival]
 
         result
       end