+++ app/builders/dnd2024_character/classes/bard_builder.rb
@@ -9,8 +9,9 @@ module Dnd2024Character
       def call(result:)
         result[:weapon_core_skills] = result[:weapon_core_skills].concat(WEAPON_CORE).uniq
         result[:armor_proficiency] = result[:armor_proficiency].concat(ARMOR).uniq
-        result[:abilities] = { str: 10, dex: 14, con: 11, int: 12, wis: 13, cha: 15 }
-        result[:health] = { current: 8, max: 8, temp: 0 }
+        result[:abilities] = { str: 8, dex: 14, con: 12, int: 13, wis: 10, cha: 15 }
+        result[:health] = { current: 9, max: 9, temp: 0 }
+        result[:any_skill_boosts] += 3
 
         result
       end