+++ app/builders/dnd2024_character/classes/warlock_builder.rb
@@ -9,8 +9,10 @@ module Dnd2024Character
       def call(result:)
         result[:weapon_core_skills] = result[:weapon_core_skills].concat(WEAPON_CORE).uniq
         result[:armor_proficiency] = result[:armor_proficiency].concat(ARMOR).uniq
-        result[:abilities] = { str: 10, dex: 12, con: 11, int: 13, wis: 15, cha: 14 }
-        result[:health] = { current: 8, max: 8, temp: 0 }
+        result[:abilities] = { str: 8, dex: 14, con: 13, int: 12, wis: 10, cha: 15 }
+        result[:health] = { current: 9, max: 9, temp: 0 }
+        result[:skill_boosts] += 2
+        result[:skill_boosts_list] = %w[arcana deception history intimidation investigation nature religion]
 
         result
       end