+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - rendering spell stats for DC20 characters
 - selecting talents for DC20 characters
+- guide mode for D&D 2024 characters
 
 ### Modified
 - leveling for DH characters